backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pandas
pandas_ta
python-dotenv
requests
apscheduler
numpy<2.0.0
psycopg2-binary
